added a test for hssp conversion
