+++ src/hssp-convert-3to1.cpp
@@ -33,7 +33,8 @@
 #include <boost/regex.hpp>
 
 #include <cmath>
-#include <iostream>
+
+#include "hssp-convert-3to1.h"
 
 #if P_WIN
 #pragma warning (disable: 4267)
@@ -45,8 +46,6 @@ namespace ba = boost::algorithm;
 namespace io = boost::iostreams;
 namespace fs = boost::filesystem;
 namespace po = boost::program_options;
-
-int VERBOSE = 0;
 
 // --------------------------------------------------------------------
 // basic named sequence type and a multiple sequence alignment container
@@ -1079,109 +1078,4 @@ void ConvertHsspFile(std::istream& in, std::ostream& out)
   CreateHSSPOutput(id, header, 0.05f, seqlength, nchain, kchain,
                    ba::join(usedChains, ", "), msa,
     hits, residues, out);
-}
-
-// --------------------------------------------------------------------
-
-int main(int argc, char* const argv[])
-{
-  try
-  {
-    po::options_description desc("hsspconv options");
-    desc.add_options()
-      ("help,h", "Display help message")
-      ("input,i", po::value<std::string>(), "Input PDB file (or PDB ID)")
-      ("output,o",
-       po::value<std::string>(),
-       "Output file, use 'stdout' to output to screen")
-      ("version", "Show version number");
-
-    po::positional_options_description p;
-    p.add("input", 1);
-    p.add("output", 2);
-
-    po::variables_map vm;
-    po::store(po::command_line_parser(
-          argc, argv).options(desc).positional(p).run(), vm);
-
-    //fs::path home = get_home();
-    //if (fs::exists(home / ".mkhssprc"))
-    //{
-    //  fs::ifstream rc(home / ".mkhssprc");
-    //  po::store(po::parse_config_file(rc, desc), vm);
-    //}
-
-    po::notify(vm);
-
-    if(vm.count("version")>0)
-    {
-      std::cout << "hssp converter version "<<XSSP_VERSION << std::endl;
-      exit(0);
-    }
-
-    if (vm.count("help"))
-    {
-      std::cerr << desc << std::endl;
-      exit(1);
-    }
-
-    io::filtering_stream<io::input> in;
-    fs::ifstream ifs;
-
-    if (vm.count("input") == 0)
-      in.push(std::cin);
-    else
-    {
-      fs::path input = vm["input"].as<std::string>();
-      ifs.open(input, std::ios::binary);
-
-      if (not ifs.is_open())
-        throw mas_exception(
-            boost::format("Could not open input file '%s'") % input);
-
-#ifdef HAVE_LIBBZ2
-      if (input.extension() == ".bz2")
-        in.push(io::bzip2_decompressor());
-      else if (input.extension() == ".gz")
-        in.push(io::gzip_decompressor());
-#endif
-      in.push(ifs);
-    }
-
-    if (vm.count("output") == 0)
-      ConvertHsspFile(in, std::cout);
-    else
-    {
-      fs::path output = vm["output"].as<std::string>();
-
-      fs::ofstream ofs(output, std::ios::binary);
-      if (not ofs.is_open())
-        throw mas_exception(
-            boost::format("Could not open output file '%s'") % output);
-
-      io::filtering_stream<io::output> out;
-
-#ifdef HAVE_LIBBZ2
-      if (output.extension() == ".bz2")
-        out.push(io::bzip2_compressor());
-      else if (output.extension() == ".gz")
-        out.push(io::gzip_compressor());
-#endif
-      out.push(ofs);
-
-      ConvertHsspFile(in, out);
-    }
-  }
-  catch (const std::exception& e)
-  {
-    std::cerr << e.what() << std::endl;
-    exit(1);
-  }
-  catch (...)
-  {
-    std::cerr << "Unknown exception" << std::endl;
-    exit(1);
-  }
-
-  return 0;
 }
