+++ src/hssp-convert-3to1.h
@@ -1,0 +1,8 @@
+#ifndef HSSP_CONVERT_3TO1_H
+#define HSSP_CONVERT_3TO1_H
+
+#include <iostream>
+
+void ConvertHsspFile(std::istream& in, std::ostream& out);
+
+#endif // HSSP_CONVERT_3TO1_H
